Set secret permissions at runtime (for podman < 4.8.0 compatibility)
Prior to podman 4.8.0, the defaultMode for Secret type volumes was not
used. This causes SSH key data to have incorrect permissions when
mounted into containers.

This commit updates the fix_ssh_perms.sh script for the brics_slurm
container to correctly set permissions on container start. A similar
script is added to brics_jupyterhub container and run as part of the
CMD on container start. The dev_dummyauth base pod.yaml is modified
to make the SSH key data not read only when mounted into containers
to allow runtime permission modification.